requirements.txt
alembic==1.7.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
dateutils==0.6.12
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2==2.9.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.26
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
